+++ Engine/Source/Runtime/Platform/Windows/WindowsWindow.h
@@ -11,15 +11,15 @@ namespace Wi
 	public:
 		explicit WindowsWindow(const WindowDefinition& definition, HINSTANCE instance);
 		void Destroy() override;
-		i32 GetWidth() override;
-		i32 GetHeight() override;
+		u16 GetWidth() override;
+		u16 GetHeight() override;
 
 		static const wchar_t* AppWindowClass;
 
 	private:
 		HWND m_WindowHandle = nullptr;
-		i32 m_Width;
-		i32 m_Height;
+		int m_Width;
+		int m_Height;
 	};
 }
 #endif
